pyproject.toml
[project]
name = "mesa-de-ayuda-comunicarlos"
version = "0.1.0"
description = "Mesa de Ayuda - Cooperativa Comunicarlos - PP5"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "sqlalchemy>=2.0",
]

[dependency-groups]
dev = ["pytest>=8.0", "httpx>=0.28"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
